Move SignatureAddForm to containers
