+++ src/components/theme-legacy/petition.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import { text2paraJsx } from '../../lib.js'
-import SignatureAddForm from '../signature-add-form.js'
+import SignatureAddForm from '../../containers/signature-add-form.js'
 import SignatureCount from '../signature-count.js'
 import SignatureList from '../signature-list.js'
 import PetitionFlagForm from '../petition-flag-form.js'
